fix: split mock-worker config out of agents_config.go for backend-size gate
Move mock-worker types and parsing into pkg/config/mock_workers_config.go so
agents_config.go stays under the 1000-line CI limit.
